requirements.txt
# JobForge Dependencies

# HTTP client
httpx>=0.25.0

# Browser automation
playwright>=1.40.0

# YAML parsing
pyyaml>=6.0

# Resume parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# CLI utilities
rich>=13.0.0
click>=8.0.0

# Data handling
pandas>=2.0.0

# Type hints
typing-extensions>=4.0.0
